fix(NAS-MPI): change pcvs recipe and update to 3.4.3
This new version of the NAS Parallel Benchmarks adds the following:
- fortran rework to use the MPI module instead of the old mpif.h header
- compile number of processes agnostic binaries

It means that the recipe had to be refactored to reflect the new
binary names. Furthermore, the recipe had wrong constraints on the
number of processes one test could used. It has been changed to avoid
errors related to the number of processes requested.