fix(OSU): fix pcvs.setup and update to 7.5.1
The new version adds tests to other languages and changes the layout of
the repository. The pcvs setup has been updated to follow new layout.

New tests come with this new version. None of them have been added to
the `pcvs.setup`. It can be a good idea to split the top level
`pcvs.setup` in smaller ones to run only part of the OSU suite.